Add inventory management with EF Core and unit of work
Implemented inventory repository and unit of work patterns using EF Core. Added methods for inventory queries, paging, and transaction management. Refactored DTOs to use decimals and shared base classes. Introduced argument validation extension. Implemented request handlers for inventory operations with business logic. Updated DI registration and removed obsolete code.

+++ DeepCore/DeepCore/RequestHandlers/Inventories/AdjustInventoryRequest.cs
@@ -3,7 +3,7 @@ namespace DeepCore.RequestHandlers.Inventories
     public class AdjustInventoryRequest : IRequest<AdjustInventoryResponse>
     {
         public long ProductId { get; set; }
-        public int QuantityDifference { get; set; }
+        public decimal QuantityDifference { get; set; }
         public string? Reason { get; set; }
     }
 }
